refactor: rename Cart class to CartItem
